Move InstancesTableView to own file
